TLCPM-494 change /bulkUpload API call to return status quickly, while leaving the actual processing of the requested site into back running threads
